feat: DRY'd up code by removing unused algo

+++ src/main/java/ca/mcmaster/se2aa4/island/team121/businessdrivenobjects/Point.java
@@ -1,7 +1,7 @@
 package ca.mcmaster.se2aa4.island.team121.businessdrivenobjects;
 
-public record Point(int x, int y) {
-    public Double getDistance(Point other) {
+public record Point(double x, double y) {
+    public double getDistance(Point other) {
         return Math.sqrt(Math.pow(this.x() - other.x(), 2) + Math.pow(this.y() - other.y(), 2));
     }
 }
